% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_ogr.R
\name{as_ogr}
\alias{as_ogr}
\alias{as_ogr.character}
\alias{as_ogr.OGRSQLConnection}
\title{Delayed read for vector resources}
\usage{
as_ogr(x, layer, ..., query = NA, ignore_lyrs = "gpkg_|rtree_|sqlite_")

\method{as_ogr}{character}(x, layer, ..., query = NA, ignore_lyrs = "gpkg_|rtree_|sqlite_")

\method{as_ogr}{OGRSQLConnection}(x, layer, ..., query = NA, ignore_lyrs = "gpkg_|rtree_|sqlite_")
}
\arguments{
\item{x}{the data source (file path, url, or database connection)}

\item{layer}{layer name (varies by driver, may be a file name without
extension); in case \code{layer} is missing, \code{st_read} will read the
first layer of \code{dsn}, give a warning and (unless \code{quiet = TRUE})
print a message when there are multiple layers, or give an error if there
are no layers in \code{dsn}. If \code{dsn} is a database connection, then
\code{layer} can be a table name or a database identifier (see
\code{\link[DBI]{Id}}). It is also possible to omit \code{layer} and rather
use the \code{query} argument.}

\item{...}{parameter(s) passed on to \link[sf]{st_as_sf}}

\item{query}{SQL query to pass in directly}

\item{ignore_lyrs}{pattern for layers to be ignored description}
}
\value{
a 'tbl_OGRSQLConnection'
}
\description{
A lazy data frame for GDAL vector data sources. as_ogr is DBI compatible and designed to work with dplyr.
}
\details{
The output of `as_ogr()` is a 'tbl_OGRSQLConnection` that extends `tbl_dbi` and
may be used with functions and workflows in the normal DBI way, see [OGRSQL()] for
the as_ogr DBI support.

To obtain an in memory data frame use an explict `collect()` or `st_as_sf()`.
A call to `collect()` is triggered by `st_as_sf()` and will add the sf class
to the output.
}
