% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subset.R
\name{extract_gpkg_data}
\alias{extract_gpkg_data}
\title{Extract Data from Arrow Stores}
\usage{
extract_gpkg_data(gpkg, vpu, ids, lyrs, outfile = NULL)
}
\arguments{
\item{lyrs}{layers to extract}

\item{outfile}{If gpkg file path is provided, data will be written to a file.}

\item{hook}{a local or s3 hydrofabric directory}
}
\value{
list or file path
}
\description{
Extract Data from Arrow Stores
}
