% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_origin.R
\name{find_origin}
\alias{find_origin}
\title{Find an origin from indexed IDs}
\usage{
find_origin(
  network,
  id,
  type = c("id", "comid", "hl_uri", "poi_id", "nldi_feature", "xy")
)
}
\arguments{
\item{network}{A `dplyr`-compatible object.}

\item{id}{A queryable identifier of type `type`.}

\item{type}{An index type describing `id`.}
}
\value{
A network origin. If a single origin is not found,
         then an exception is raised.
}
\description{
Find an origin from indexed IDs
}
