% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_origin.R
\name{find_origin_query}
\alias{find_origin_query}
\title{S3 dispatch on query identifier type}
\usage{
find_origin_query(id, network)
}
\arguments{
\item{id}{A queryable identifier, see `find_origin`.}

\item{network}{A `dplyr`-compatible object.}
}
\value{
`network` after applying a [dplyr::filter] expression.
}
\description{
S3 dispatch on query identifier type
}
\keyword{internal}
