% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_origin.R
\name{findOriginQuery}
\alias{findOriginQuery}
\title{S3 method for dispatching on query type}
\usage{
findOriginQuery(id, network, ...)
}
\value{
Arrow Table/Deferred connection
}
\description{
S3 method for dispatching on query type
}
\keyword{internal}
