% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_origin_gpkg.R
\name{findOriginQueryGPKG}
\alias{findOriginQueryGPKG}
\title{S3 method for dispatching on query type}
\usage{
findOriginQueryGPKG(id, gpkg, ...)
}
\value{
Arrow Table/Deferred connection
}
\description{
S3 method for dispatching on query type
}
\keyword{internal}
