% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hydrofabric}
\alias{get_hydrofabric}
\title{Download a Hydrofabric Geopackage}
\usage{
get_hydrofabric(
  url = "https://lynker-spatial.s3-us-west-2.amazonaws.com/hydrofabric",
  version = "2.2",
  domain = "conus",
  type = "nextgen",
  outfile = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{url}{A character string specifying the base URL of the hydrofabric repository. Defaults to `'https://lynker-spatial.s3-us-west-2.amazonaws.com/hydrofabric'`.}

\item{version}{A character string indicating the version of the hydrofabric to download. Defaults to `'2.2'`.}

\item{domain}{A character string specifying the geographic domain of the hydrofabric. Defaults to `'conus'`.}

\item{type}{A character string indicating the type of hydrofabric. Defaults to `'nextgen'`.}

\item{outfile}{A character string specifying the path to save the downloaded file. If `NULL`, the file will not be saved. Defaults to `NULL`.}

\item{overwrite}{A logical value indicating whether to overwrite an existing file. Defaults to `FALSE`.}
}
\value{
The function returns the path to the downloaded file (`outfile`).
}
\description{
Downloads a hydrofabric Geopackage from a specified URL and saves it to a local file.
}
\examples{
\dontrun{
# Download the default hydrofabric file
get_hydrofabric(outfile = "conus_nextgen.gpkg")

# Specify a different domain and version
get_hydrofabric(
  version = "3.0",
  domain = "hawaii",
  outfile = "hawaii_nextgen.gpkg",
  overwrite = TRUE
)
}
}
