% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shortest_path.R
\name{get_shortest_path}
\alias{get_shortest_path}
\title{Find the shortest path between two points in a hydrofabric network}
\usage{
get_shortest_path(
  start_id,
  end_id,
  gpkg,
  filename = NULL,
  lyrs = c("divides", "flowpaths", "network", "nexus", "flowpath-attributes")
)
}
\arguments{
\item{start_id}{the starting NHDPlusV2 COMID id of the shortest path. datatype: int / vector of int e.g., 61297116 or c(61297116 , 6129261)}

\item{end_id}{the ending NHDPlusV2 COMID id of the shortest path. datatype: int / vector of int e.g., 61297116 or c(61297116 , 6129261)}

\item{gpkg}{a local gpkg file}

\item{filename}{If filename is provided, data will be written using the filename}

\item{lyrs}{layers to extract}
}
\value{
An sf object containing the shortest path
}
\description{
Find the shortest path between two points in a hydrofabric network
}
\details{
This function identifies the shortest path between two nodes in a hydrofabric network.
}
\author{
Caleb Novinger  <Caleb.D.Novinger@rtx.com>

Tadd Bindas <Tadd.N.Bindas@rtx.com>
}
