% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpu_gpkg.R
\name{get_vpu_fabric}
\alias{get_vpu_fabric}
\title{Extract VPU GPKG from Lynker-Spatial
Data from a Parquet Dataset will be extracted for a VPU. 
Optionally this data can be written to a GPKG.}
\usage{
get_vpu_fabric(
  vpu = "01",
  type = "reference",
  hf_version = "2.2",
  source = "s3://lynker-spatial/hydrofabric",
  outfile = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{vpu}{VPU ID}

\item{type}{hydrofabric type}

\item{hf_version}{hydrofabric version}

\item{source}{hydrofabric source (local root directory or s3 link)}

\item{outfile}{If gpkg file path is provided, data will be written to a file.}

\item{overwrite}{overwrite existing outfile file path. Default is FALSE}
}
\value{
path or list
}
\description{
Extract VPU GPKG from Lynker-Spatial
Data from a Parquet Dataset will be extracted for a VPU. 
Optionally this data can be written to a GPKG.
}
