% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfsubsetR-package.R
\docType{package}
\name{hfsubsetR-package}
\alias{hfsubsetR}
\alias{hfsubsetR-package}
\title{hfsubsetR: Hydrofabric Subsetter}
\description{
Subset Hydrofabric Data in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lynker-spatial/hfsubsetR}
  \item Report bugs at \url{https://github.com/lynker-spatial/hfsubsetR}
}

}
\author{
\strong{Maintainer}: Mike Johnson \email{jjohnson@lynker.com}

Other contributors:
\itemize{
  \item Justin Singh-Mohudpur [contributor]
}

}
\keyword{internal}
