% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_source.R
\name{is_query_source}
\alias{is_query_source}
\title{Check if an object is an hfsubset query source}
\usage{
is_query_source(query_source)
}
\arguments{
\item{query_source}{An object}
}
\value{
TRUE if `query_source` inherits from 'hfsubset_query_source'.
}
\description{
Check if an object is an hfsubset query source
}
\keyword{internal}
