% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_ogr.R
\name{OGRSQL}
\alias{OGRSQL}
\title{OGRSQL
OGRSQL driver, use to [dbConnect()] to a data source readable by sf}
\usage{
OGRSQL()
}
\description{
OGRSQL
OGRSQL driver, use to [dbConnect()] to a data source readable by sf
}
