% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGRSQLConnection.R, R/sf_ogr.R
\docType{class}
\name{OGRSQLConnection-class}
\alias{OGRSQLConnection-class}
\alias{show,OGRSQLConnection-method}
\alias{dbSendQuery,OGRSQLConnection,ANY-method}
\alias{dbReadTable,OGRSQLConnection,character-method}
\alias{dbListTables,OGRSQLConnection-method}
\alias{dbExistsTable,OGRSQLConnection,ANY-method}
\alias{dbDisconnect,OGRSQLConnection-method}
\title{Class OGRSQLConnection (and methods)}
\usage{
\S4method{show}{OGRSQLConnection}(object)

\S4method{dbSendQuery}{OGRSQLConnection,ANY}(conn, statement, ...)

\S4method{dbReadTable}{OGRSQLConnection,character}(conn, name, ...)

\S4method{dbListTables}{OGRSQLConnection}(conn, ...)

\S4method{dbExistsTable}{OGRSQLConnection,ANY}(conn, name, ...)

\S4method{dbDisconnect}{OGRSQLConnection}(conn, ...)
}
\description{
OGRSQLConnection objects are created by passing [OGRSQL()] as first
argument to [DBI::dbConnect()].
They are a superclass of the [DBIConnection-class] class.
}
\keyword{internal}
