% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGRSQLDriver.R
\docType{class}
\name{OGRSQLDriver-class}
\alias{OGRSQLDriver-class}
\alias{dbDataType,OGRSQLDriver,ANY-method}
\alias{dbIsValid,OGRSQLDriver-method}
\alias{dbUnloadDriver,OGRSQLDriver-method}
\alias{dbGetInfo,OGRSQLDriver-method}
\title{Class OGRSQLDriver}
\usage{
\S4method{dbDataType}{OGRSQLDriver,ANY}(dbObj, obj, ...)

\S4method{dbIsValid}{OGRSQLDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{OGRSQLDriver}(drv, ...)

\S4method{dbGetInfo}{OGRSQLDriver}(dbObj, ...)
}
\description{
OGRSQLDriver objects are created by [OGRSQL()] and used to select the correct
method in [dbConnect()].
They are a superclass of the [DBIDriver-class] class, and used purely for dispatch.
}
\keyword{internal}
