% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGRSQLResult.R
\docType{class}
\name{OGRSQLResult-class}
\alias{OGRSQLResult-class}
\alias{show,OGRSQLResult-method}
\alias{dbFetch,OGRSQLResult-method}
\alias{dbClearResult,OGRSQLResult-method}
\alias{dbHasCompleted,OGRSQLResult-method}
\title{Class OGRSQLResult (and methods)}
\usage{
\S4method{show}{OGRSQLResult}(object)

\S4method{dbFetch}{OGRSQLResult}(res, n = -1, ...)

\S4method{dbClearResult}{OGRSQLResult}(res, ...)

\S4method{dbHasCompleted}{OGRSQLResult}(res, ...)
}
\description{
OGRSQLResult objects are created by [dbSendQuery()] or [dbSendStatement()],
and encapsulate the result of an SQL statement.
They are a superclass of the [DBIResult-class] class.
The "Usage" section lists the class methods overridden by \pkg{lazsf}.
}
\seealso{
The corresponding generic functions
[DBI::dbFetch()], [DBI::dbClearResult()], and
[DBI::dbHasCompleted()].
}
\keyword{internal}
