% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_subset.R
\name{query_extract}
\alias{query_extract}
\title{Perform data extraction from a query}
\usage{
query_extract(query)
}
\arguments{
\item{query}{A `hfsubset_query` object}
}
\value{
A list of hydrofabric layers, or the path to the sink of the query
}
\description{
Perform data extraction from a query
}
\note{
This should be called from query_subset().
}
\keyword{internal}
