% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_get_layers}
\alias{query_get_layers}
\title{Accessor to get the layers of a query}
\usage{
query_get_layers(query)
}
\arguments{
\item{query}{A `hfsubset_query` object}
}
\description{
Accessor to get the layers of a query
}
\keyword{internal}
