% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_get_sink}
\alias{query_get_sink}
\title{Accessor to get the sink of a query}
\usage{
query_get_sink(query)
}
\arguments{
\item{query}{A `hfsubset_query` object}
}
\description{
Accessor to get the sink of a query
}
\keyword{internal}
