% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_set_layers}
\alias{query_set_layers}
\title{Set the layers of a query}
\usage{
query_set_layers(query, layers)
}
\arguments{
\item{query}{A `hfsubset_query` object}

\item{layers}{A `character` vector of layer names}
}
\value{
`query` with the layers included
}
\description{
Set the layers of a query
}
