% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_set_sink}
\alias{query_set_sink}
\title{Set the sink of a query}
\usage{
query_set_sink(query, sink, overwrite = FALSE)
}
\arguments{
\item{query}{A `hfsubset_query` object}

\item{sink}{A character path to sink}

\item{overwrite}{If TRUE, then if the sink exists, it should be overwritten}
}
\value{
`query` with the sink included
}
\description{
Set the sink of a query
}
