% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_set_source}
\alias{query_set_source}
\title{Set the source of a query}
\usage{
query_set_source(query, src)
}
\arguments{
\item{query}{A `hfsubset_query` object}

\item{src}{A `hfsubset_query_source` object}
}
\value{
`query` with the source included
}
\description{
Set the source of a query
}
\seealso{
query_source_arrow query_source_sf
}
