% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_source_arrow.R
\name{query_source_arrow}
\alias{query_source_arrow}
\title{Create a new `arrow` query source.}
\usage{
query_source_arrow(srcname, ...)
}
\arguments{
\item{srcname}{URI to `arrow`-compatible dataset}

\item{...}{Unused}
}
\value{
An `hfsubset_query_source_arrow` object
}
\description{
Create a new `arrow` query source.
}
