% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_source.R
\name{query_source_layer}
\alias{query_source_layer}
\title{Read/open a layer from a query source}
\usage{
query_source_layer(query_source, layer, ...)
}
\description{
Read/open a layer from a query source
}
\keyword{internal}
