% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_source_sf.R
\name{query_source_sf}
\alias{query_source_sf}
\title{Create a new `sf` query source.}
\usage{
query_source_sf(srcname, ...)
}
\arguments{
\item{srcname}{Path or VSI URI to source}

\item{...}{Unused}
}
\value{
An `hfsubset_query_source_sf` object
}
\description{
Create a new `sf` query source.
}
