% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_subset.R
\name{query_subset}
\alias{query_subset}
\title{Execute a query subset}
\usage{
query_subset(query)
}
\arguments{
\item{query}{A `hfsubset_query` object}
}
\value{
A list of hydrofabric layers, or the path to the sink of the query
}
\description{
Execute a query subset
}
