% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfsubsetR-package.R
\docType{package}
\name{hfsubsetR-package}
\alias{hfsubsetR}
\alias{hfsubsetR-package}
\title{hfsubsetR: Hydrofabric Subsetter}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Subset Hydrofabric Data in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/owp-spatial/hfsubsetR}
  \item Report bugs at \url{https://github.com/owp-spatial/hfsubsetR}
}

}
\author{
\strong{Maintainer}: Mike Johnson \email{mike.johnson@noaa.gov}

Authors:
\itemize{
  \item Justin Singh-Mohudpur
}

}
\keyword{internal}
